+++ database/src/main/java/com/coderizzard/database/data/database/AppDatabase.kt
@@ -4,6 +4,8 @@ import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
 import com.coderizzard.database.data.DbConverter
+import com.coderizzard.database.data.database.dao.QuestionDao
+import com.coderizzard.database.data.database.dao.QuizDao
 import com.coderizzard.database.data.database.model.QuizEntity
 import com.coderizzard.database.data.database.model.question.IdentificationQuestionEntity
 import com.coderizzard.database.data.database.model.question.MCOptionEntity
