+++ database/src/main/java/com/coderizzard/database/data/di/DatabaseModule.kt
@@ -3,8 +3,8 @@ package com.coderizzard.database.data.di
 import android.content.Context
 import androidx.room.Room
 import com.coderizzard.database.data.database.AppDatabase
-import com.coderizzard.database.data.database.QuestionDao
-import com.coderizzard.database.data.database.QuizDao
+import com.coderizzard.database.data.database.dao.QuestionDao
+import com.coderizzard.database.data.database.dao.QuizDao
 import com.coderizzard.database.data.repository.QuestionRepositoryImpl
 import com.coderizzard.database.data.repository.QuizRepositoryImpl
 import com.coderizzard.database.domain.repository.QuestionRepository
