+++ database/src/main/java/com/coderizzard/database/data/repository/QuestionRepositoryImpl.kt
@@ -1,9 +1,8 @@
 package com.coderizzard.database.data.repository
 
 import com.coderizzard.core.data.model.question.Question
-import com.coderizzard.database.data.database.QuestionDao
+import com.coderizzard.database.data.database.dao.QuestionDao
 import com.coderizzard.database.data.database.model.question.IdentificationQuestionEntity
-import com.coderizzard.database.data.database.model.question.MCOptionEntity
 import com.coderizzard.database.data.database.model.question.MCQuestionEntity
 import com.coderizzard.database.data.database.model.question.QuestionEntity
 import com.coderizzard.database.domain.repository.QuestionRepository
